+++ Soggins/FrogproofManager.cs
@@ -1,12 +1,5 @@
 ﻿using HarmonyLib;
-using Microsoft.Extensions.Logging;
-using Nanoray.Shrike;
-using Nanoray.Shrike.Harmony;
 using Shockah.Shared;
-using System;
-using System.Collections.Generic;
-using System.Reflection;
-using System.Reflection.Emit;
 
 namespace Shockah.Soggins;
 
@@ -23,16 +16,6 @@ public sealed class FrogproofManager : HookManager<IFrogproofHook>
 
 	internal static void ApplyPatches(Harmony harmony)
 	{
-		harmony.TryPatch(
-			logger: Instance.Logger!,
-			original: () => AccessTools.DeclaredMethod(typeof(Card), nameof(Card.Render)),
-			transpiler: new HarmonyMethod(typeof(FrogproofManager), nameof(Card_Render_Transpiler))
-		);
-		harmony.TryPatch(
-			logger: Instance.Logger!,
-			original: () => AccessTools.DeclaredMethod(typeof(Card), nameof(Card.GetAllTooltips)),
-			postfix: new HarmonyMethod(typeof(FrogproofManager), nameof(Card_GetAllTooltips_Postfix))
-		);
 		harmony.TryPatch(
 			logger: Instance.Logger!,
 			original: () => AccessTools.DeclaredMethod(typeof(Ship), nameof(Ship.Set)),
@@ -74,91 +57,6 @@ public sealed class FrogproofManager : HookManager<IFrogproofHook>
 		return null;
 	}
 
-	private static IEnumerable<CodeInstruction> Card_Render_Transpiler(IEnumerable<CodeInstruction> instructions, MethodBase originalMethod)
-	{
-		try
-		{
-			return new SequenceBlockMatcher<CodeInstruction>(instructions)
-				.Find(
-					ILMatches.Ldloc<CardData>(originalMethod).ExtractLabels(out var labels).Anchor(out var findAnchor),
-					ILMatches.Ldfld("buoyant"),
-					ILMatches.Brfalse
-				)
-				.Find(
-					ILMatches.Ldloc<Vec>(originalMethod).CreateLdlocInstruction(out var ldlocVec),
-					ILMatches.Ldfld("y"),
-					ILMatches.LdcI4(8),
-					ILMatches.Ldloc<int>(originalMethod).CreateLdlocaInstruction(out var ldlocaCardTraitIndex),
-					ILMatches.Instruction(OpCodes.Dup),
-					ILMatches.LdcI4(1),
-					ILMatches.Instruction(OpCodes.Add),
-					ILMatches.Stloc<int>(originalMethod)
-				)
-				.Anchors().PointerMatcher(findAnchor)
-				.Insert(
-					SequenceMatcherPastBoundsDirection.Before, SequenceMatcherInsertionResultingBounds.IncludingInsertion,
-					new CodeInstruction(OpCodes.Ldarg_1).WithLabels(labels),
-					new CodeInstruction(OpCodes.Ldarg_3),
-					new CodeInstruction(OpCodes.Ldarg_0),
-					ldlocaCardTraitIndex,
-					ldlocVec,
-					new CodeInstruction(OpCodes.Call, AccessTools.DeclaredMethod(typeof(FrogproofManager), nameof(Card_Render_Transpiler_RenderFrogproofIfNeeded)))
-				)
-				.AllElements();
-		}
-		catch (Exception ex)
-		{
-			Instance.Logger!.LogError("Could not patch method {Method} - {Mod} probably won't work.\nReason: {Exception}", originalMethod, Instance.Name, ex);
-			return instructions;
-		}
-	}
-
-	private static void Card_Render_Transpiler_RenderFrogproofIfNeeded(G g, State? state, Card card, ref int cardTraitIndex, Vec vec)
-	{
-		state ??= g.state;
-
-		var frogproofType = Instance.FrogproofManager.GetFrogproofType(state, state.route as Combat, card, FrogproofHookContext.Rendering);
-		if (frogproofType == FrogproofType.None)
-			return;
-		// state is usually DB.fakeState here; use g.state instead
-		if (frogproofType == FrogproofType.InnateHiddenIfNotNeeded && (g.state.IsOutsideRun() || !Instance.Api.IsRunWithSmug(g.state)))
-			return;
-		Draw.Sprite((Spr)Instance.FrogproofSprite.Id!.Value, vec.x, vec.y - 8 * cardTraitIndex++);
-	}
-
-	private static void Card_GetAllTooltips_Postfix(Card __instance, State s, bool showCardTraits, ref IEnumerable<Tooltip> __result)
-	{
-		if (!showCardTraits)
-			return;
-
-		var frogproofType = Instance.FrogproofManager.GetFrogproofType(s, s.route as Combat, __instance, FrogproofHookContext.Rendering);
-		if (frogproofType == FrogproofType.None)
-			return;
-		// state is usually DB.fakeState here; use StateExt.Instance instead
-		if (frogproofType == FrogproofType.InnateHiddenIfNotNeeded && ((StateExt.Instance ?? s).IsOutsideRun() || !Instance.Api.IsRunWithSmug(StateExt.Instance ?? s)))
-			return;
-
-		static IEnumerable<Tooltip> ModifyTooltips(IEnumerable<Tooltip> tooltips)
-		{
-			bool yieldedFrogproof = false;
-
-			foreach (var tooltip in tooltips)
-			{
-				if (!yieldedFrogproof && tooltip is TTGlossary glossary && glossary.key.StartsWith("cardtrait.") && glossary.key != "cardtrait.unplayable")
-				{
-					yield return Instance.Api.FrogproofCardTraitTooltip;
-					yieldedFrogproof = true;
-				}
-				yield return tooltip;
-			}
-
-			if (!yieldedFrogproof)
-				yield return Instance.Api.FrogproofCardTraitTooltip;
-		}
-
-		__result = ModifyTooltips(__result);
-	}
-
 	private static void Ship_Set_Postfix(Ship __instance, Status status, int n)
 	{
 		if (StateExt.Instance is not { } state || state.ship != __instance)
@@ -181,7 +79,7 @@ public sealed class FrogproofCardTraitFrogproofHook : IFrogproofHook
 	private FrogproofCardTraitFrogproofHook() { }
 
 	public FrogproofType? GetFrogproofType(State state, Combat? combat, Card card, FrogproofHookContext context)
-		=> card is IFrogproofCard frogproofCard && frogproofCard.IsFrogproof(state, combat) ? FrogproofType.Innate : null;
+		=> ModEntry.Instance.Helper.Content.Cards.GetCardTraitState(state, card, ModEntry.Instance.FrogproofTrait).IsActive ? FrogproofType.Innate : null;
 
 	public void PayForFrogproof(State state, Combat? combat, Card card) { }
 }