+++ Soggins/InternalInterfaces.cs
@@ -1,5 +1,7 @@
 ﻿using CobaltCoreModding.Definitions.ModContactPoints;
 using HarmonyLib;
+using Nickel;
+using System.Collections.Generic;
 
 namespace Shockah.Soggins;
 
@@ -19,8 +21,16 @@ internal interface IRegisterableCard
 	void InjectDialogue() { }
 }
 
-internal interface IFrogproofCard
+internal interface IFrogproofCard : IHasCustomCardTraits
 {
 	bool IsFrogproof(State state, Combat? combat)
 		=> true;
+
+	IReadOnlySet<ICardTraitEntry> IHasCustomCardTraits.GetInnateTraits(State state)
+	{
+		var traits = new HashSet<ICardTraitEntry>();
+		if (IsFrogproof(state, state.route as Combat))
+			traits.Add(ModEntry.Instance.FrogproofTrait);
+		return traits;
+	}
 }